pyproject.toml
[project]
name = "license-crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.44",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tomli>=2.2.1",
]
